feat: update image updater

+++ teams/app-devs/image-updater.js
@@ -6,6 +6,7 @@ const configFile = require("./config.json");
 const required_env_variables = [
   {
     configPath: ["bffApi", "dockerImageURI"],
+    kind: "Deployment",
     filePath: "./dev/templates/bff-api/app.yaml",
     keyPath: ["spec", "template", "spec", "containers", 0],
     key: "image",
@@ -14,32 +15,66 @@ const required_env_variables = [
 
 try {
   for (let toReplace of required_env_variables) {
-    const doc = yaml.load(
+    const docs = yaml.loadAll(
       fs.readFileSync(path.join(__dirname, toReplace.filePath), "utf8")
     );
 
-    let ref = doc;
-    for (let field of toReplace.keyPath) {
-      ref = ref[field];
-    }
+    console.log(docs.length);
 
-    let configRef = configFile;
+    let start = true;
 
-    for (let field of toReplace.configPath) {
-      configRef = configRef[field];
-    }
+    for (let doc of docs) {
+      let ref = doc;
 
-    console.log(configRef);
-
-    ref[toReplace.key] = configRef;
-
-    fs.writeFileSync(
-      path.join(__dirname, toReplace.filePath),
-      yaml.dump(doc, {lineWidth: -1}),
-      {
-        encoding: "utf-8",
+      if (!doc) {
+        continue;
       }
-    );
+
+      if (start) {
+        fs.writeFileSync(path.join(__dirname, toReplace.filePath), `---\n`, {
+          encoding: "utf-8",
+        });
+      } else {
+        fs.appendFileSync(path.join(__dirname, toReplace.filePath), `---\n`, {
+          encoding: "utf-8",
+        });
+      }
+
+      start = false;
+
+      if (doc["kind"] !== toReplace.kind) {
+        fs.appendFileSync(
+          path.join(__dirname, toReplace.filePath),
+          yaml.dump(doc, {lineWidth: -1}),
+          {
+            encoding: "utf-8",
+          }
+        );
+        continue;
+      }
+
+      for (let field of toReplace.keyPath) {
+        ref = ref[field];
+      }
+
+      let configRef = configFile;
+
+      for (let field of toReplace.configPath) {
+        configRef = configRef[field];
+      }
+
+      if (configRef) {
+        ref[toReplace.key] = configRef;
+      }
+
+      fs.appendFileSync(
+        path.join(__dirname, toReplace.filePath),
+        yaml.dump(doc, {lineWidth: -1}),
+        {
+          encoding: "utf-8",
+        }
+      );
+    }
   }
 } catch (e) {
   console.log(e);
